fix colum types
